requirements.txt
#Python 3.11.2
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
